Create interface with abc library in Python with a run method declaraion
