+++ DataConfiguration.py
@@ -1,4 +1,4 @@
-from DataGenerator import DataGenerator
+from RandomDataGenerator import RandomDataGenerator
 
 # This file will contrain the core configuration for running the consumer and producer.
 # It will also contain the functions needed to generate data for the different topics.
@@ -11,7 +11,10 @@ configuration = {
         },
         "Avro Schema": "prices.avsc",
         "Serializer": "Avro",
-        "Data" : DataGenerator.generate_price_entity,
+        "Data" : RandomDataGenerator.generate_price_entity,
+        "Data Args": {
+            "Type": "Price"
+        },
         "Data Queue Max Size": 40,
         "Number of Processes" : 11,
         "Number of Data Generation Processes": 1,
@@ -24,7 +27,7 @@ configuration = {
         },
         "Avro Schema": "positions.avsc",
         "Serializer": "Avro",
-        "Data" : DataGenerator.generate_position_entity,
+        "Data" : RandomDataGenerator.generate_position_entity,
         "Data Queue Max Size": 20,
         "Number of Processes" : 4,
         "Number of Data Generation Processes": 1,
@@ -37,7 +40,7 @@ configuration = {
         },
         "Avro Schema": "instrument_reference_data.avsc",
         "Serializer": "Avro",
-        "Data" : DataGenerator.generate_inst_ref_entity,
+        "Data" : RandomDataGenerator.generate_inst_ref_entity,
         "Data Queue Max Size": 5,
         "Number of Processes" : 1,
         "Number of Data Generation Processes": 1,
